add selectors for transfers / transactions; clearify the diff between the two by naming components more consistently
