+++ src/components/index.ts
@@ -21,8 +21,8 @@ import {
   SearchUsers,
   SelectedUserReferralsRow,
   SelectedUsersReferralsTable,
-  SelectedUsersTransfersTable,
-  SelectedUserTransferRow,
+  SelectedUsersTransactionsTable,
+  SelectedUserTransactionRow,
   TransactionsPagination,
   TransfersEmptyState,
   TransferSummary,
@@ -69,8 +69,8 @@ export {
   SelectedUserReferralsSection,
   SelectedUserSection,
   SelectedUsersReferralsTable,
-  SelectedUsersTransfersTable,
-  SelectedUserTransferRow,
+  SelectedUsersTransactionsTable,
+  SelectedUserTransactionRow,
   SelectedUserTransfersSection,
   SingleColumnContentWrapper,
   Spinner,
