+++ src/components/molecules/SelectedUserTransactionRow/index.ts
@@ -1,0 +1,3 @@
+import SelectedUserTransactionRow from './SelectedUserTransactionsRow';
+
+export default SelectedUserTransactionRow;
