--- src/components/molecules/SelectedUserTransferRow/index.ts
@@ -1,3 +1,0 @@
-import SelectedUserTransferRow from './SelectedUserTransferRow';
-
-export default SelectedUserTransferRow;
