+++ src/components/molecules/SelectedUsersTransactionsTable/index.ts
@@ -1,0 +1,3 @@
+import SelectedUsersTransactionsTable from './SelectedUsersTransactionsTable';
+
+export default SelectedUsersTransactionsTable;
