--- src/components/molecules/SelectedUsersTransfersTable/index.ts
@@ -1,3 +1,0 @@
-import SelectedUsersTransfersTable from './SelectedUsersTransfersTable';
-
-export default SelectedUsersTransfersTable;
