+++ src/hooks/index.ts
@@ -2,6 +2,7 @@ import useDebounce from './useDebounce';
 import useRegisterUser from './useRegisterUser';
 import useUserAccounts from './useUserAccounts';
 import useUserKYC from './useUserKYC';
+import useUserTransactionSelector from './useUserTransactionSelector';
 import useUserTransferSelector from './useUserTransferSelector';
 
 export {
@@ -9,5 +10,6 @@ export {
   useRegisterUser,
   useUserAccounts,
   useUserKYC,
+  useUserTransactionSelector,
   useUserTransferSelector,
 };
